Update kaldi_io.py
If the file is empty the reading is not proceeded starting from the next bit, the whole file is skipped starting from the failing bit. Exception catching is added in a bit dirty way, so that when there is an empty file, it is skipped and the read proceeds to the next file.